Refactor and localize System module
Moved "getName" to public in User class and added "friend class System". Simplified method descriptions, localized comments and messages, and replaced `uint` types with `int` in Date and System-related classes for consistency. Adjusted property management logic and improved memory management with `nullptr` cleanup.

+++ Client.hpp
@@ -4,7 +4,6 @@
 
 #ifndef CLIENT_HPP
 #define CLIENT_HPP
-#include <cstdint>
 #include <string>
 
 #include "User.hpp"
@@ -12,7 +11,7 @@
 
 class Client : public User {
 public:
-    void rate(const std::string &id, uint8_t rate) const;
+    void rate(const std::string &id, int rate) const;
 
     Client(const std::string &name, const std::string &password, const std::string &email);
 };
